howdy/requirements.txt
# FastAPI 서버
fastapi==0.111.0
uvicorn[standard]==0.30.1

# 데이터 모델링
pydantic==2.7.1

# 이미지 처리
Pillow==10.3.0

# HTTP 요청
requests==2.31.0

# OpenAI API 클라이언트
openai==1.30.3

# 환경 변수 로딩
python-dotenv==1.0.1

# 타입 지원 (선택적)
typing-extensions>=4.5.0
